Allow to run multiple replicas
Copy the manager code from controller runtime and change it.
With the new manager the controller will became ready only when it became leader.

most of the code is the same as the regular manager I just change the leader election function.
When a controller take the leader he also create a file under `/tmp/healthy` this way the pod will became ready.
When the pod became ready it will be added as an endpoint for the admission webhook service