update isRelatedToKubevirt to check if the pod is part of a vm object
This commit checks if the pod is related to a vm object.
Before this commit we check if the object is related to a vmi, but on the logic if kubevirt we have

vm -> vmi -> pod

but also we can have only

vmi -> pod

This update checks if the obect is only came from a vmi or it cames from a vm.

If its a vmi object allocate unique mac only, but if the pod object came form a vm object
we don't allocate it as part of the pod mac