Update pod mac structure to map of map
This commit also check if the pod name is empty.

If the pod name is empty this mean a webhook for pod was created using
the generateName field.
The pod will obtain the name by the kube-controller.

we allocate the mac address in the map but mark it ask waiting for pod.
When the pod is created by the kube-controller we get an event in the pod reconcile function.
This call allocates the pod into the podMac map and mark the mac address as allocated.