Add revert function for pod
This commit add the ability to release mac addresses that was allocated to pod if one of the allocations fails.

The allocation could fail if there is not enough free mac addresses in the pool for all the interfaces, if the user request a mac that is already allocated or invalid.