add readiness probe
Currently, manager pods turn Ready as soon as their binary starts.
However, it takes ~40 more seconds for the webhook to actually start
serving requests. In order to avoid that, we introduce a readiness
probe that marks the pod as Ready only once the server starts.

Please note that there is no way to do it properly with the current
controller-runtime. We don't have a hook point that would get executed
once the webhook has started. Therefore we have a Sleep there. That
should be dropped once
https://github.com/kubernetes-sigs/cluster-api/issues/1855
is implemented.

Signed-off-by: Petr Horacek <phoracek@redhat.com>