Add Eventually to finalizers test
The finalizers update happens in the reconcile loop
in the kubemacpool controller.
If during the finalizers-update, the object (the vm in this case)
was updated by another controller and a newer version of the object
was inserted to the database, the finalizers update will fail,
as it appears in the issue description,
and a new update attempt will be performed in the next reconcile iteration.
This will continue until the update will succeed and the finalizers will be updated.
since this update can take time we should wait for some reasonable time before failing the test.

Signed-off-by: alonSadan <asadan@redhat.com>