Implement mac range dividing to mac range hards
Adding a new type MacRangeShards, aggregates all the required properties to represent a mac-range shard.

The pod index is extracted from the pod name, (pod manager deployed with StatefulSet so we have consist pod names with index 0..N)
In order to make the pod index extracting flexible in case that something will change in the future,
it is necessary to pass "podNameIndexExtractor", of type “func (string) (string, error)”, to parsePodNameIndexInt64 function.

Next the function "calculateMacRangeShard" calculates the mac-range-shard the current pod should manage without overlapping,
the calculation is based on the pod index so each pod will manage the same mac-range-shard for a given mac range.

Signed-off-by: Or Mergi <ormergi@redhat.com>