Remove 'Expect' for spesific error messages
In this test we expect for a sertain error message from
the virtClient when creating a VM, this message is changed
and no longer includes 'every network must be mapped to an interface'.

Since we expect to the VM creating to fail due to not valid
request we dont need to look for spesific error message,
we will get an error from kubevirt validition webhook.

Signed-off-by: Or Mergi <ormergi@redhat.com>