Change leadership mechanism (#182)
* remove old  leader election mechanism.

Currently we use an external leader election where only the elected pod actually runs the controller-runtime manager.
We want to change this behavior so that all pods run the controller-runtime manager, thus are shown as ready.
To do that we remove the od code of the leader election.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* add mgr to KubeMacPoolManager and use mgr.Elected() to mark only the elected leader pod
This also means we get can remove the use of leaderElectionChannel

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* add retryOnConflict to better handle future conflict in markPodAsLeader()

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* move leader election into the controller-runtime manager.

in thie commit we move waitToStartLeading() to a separate go routine, so that both pods will start the controller-runtime manager.
we then add the leader election inside the controller runtime manager.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* adding a new readinessGate status to kubemacpool pods.

since the webhook service operates only on pods with kubemacpool leader label, we cant start the webhooks until the election is complete and a leader as been chosen.
in order to do that, we add a new status to the pod, to prevent the pods to be ready and the webhhoks to start until the label is set.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* adding the new readinessGate status to the manifest. also adding the appropriate rbacs.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* changing sync script to wait for deployment ready

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* move pool-manager starting routines to waitToStartLeading so that they will initiate after teh election is complete

since we don't want both pods to perform pool-manager related operations, we start the pool-manager routines  only after the kubemacpool leader is selected.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* change restartKubemacpoolManagerPods in tests to reboot pods via deployment replicas, as well as making the needed changes in test to wait for both kubemacpool instead of only to the leader.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* fix code reviews

Signed-off-by: Ram Lavi <ralavi@redhat.com>