fix race between webhook and controller
Currently, it's possible that the a webhook 1 will add a mac and wrongly re-insert it to the cache on controller reconcile if it was bypassed by a new webhook 2 deleting the macs. This would cause the belated controller to re-add the mac.
To fix it we add a check that the mac is not already deleted in the controller.

Signed-off-by: Ram Lavi <ralavi@redhat.com>