remove setRangeInRangeConfigMap from tier1 tests
Currently during each start of kubemacpool tier1 test, we perform an initilazaion of the mac-range configmap.
This ability is not supported outside of the kubemacpool standalone repo, and so it makes it a problem to run tier1 tests on CNAO (for example).
We would like to change the test behavior to stop using range change.
In order to do that, we will remove the mac range change, and instead change tests the check depleted mac pools, to allocate filler vms that will deplete the mac pool. Also, we will change the range in kubemacpool repo to 16 macs in the test manifest.
This way, if some external repo like CNAO wants to run teir1 tests, they can will skip these tests using the common phrase "mac-pool is full".

Signed-off-by: Ram Lavi <ralavi@redhat.com>