change kmp manifest to allow opt-in/out for webhooks
Currently, we have 2 mutating webhooks: mutatepods and mutatevirtualmachines.
Both work in opt-in mode, meaning kmp gives service only to pods/vms reside in namespace
that were specifically opted-in for it.

Lets allow more flexibility to the user:
for each webhook two kustomize patches were added in the config/default folder (opt-in/out).
The user can choose the desired opt-mode of each webhook by copying
the patch to the appropriate opt-mode patch in the release/test folder.

The default config will be:
* release manifests: opt-out for vms, and opt-in for pods.
* test manifests: opt-out for both webhooks.

Signed-off-by: Ram Lavi <ralavi@redhat.com>