Drop the number of replicas to 1 and remove leader
We kept two replicas so there is one KMP manager running in the
background ready to take over the leadership when the original leader
dies. That should roughly lower the downtime on master failure from 4 to
2 minutes. However, we faced countless issues with the leadership
mechanism and thus decided to keep only a single replica. We should then
work on profiling and figure out what is taking so much time on the
startup without preemptive optimization.

Signed-off-by: Petr Horacek <phoracek@redhat.com>