fix cert test to use manager namespace (#240)
* fix cert test to use manager namespace

Currently the cert test uses specific manager namespace
that will only be correct if running on kubemacpool repo.
Let fix by using managerNamespace which gets the real namespace
from the cluster at  the beginning of the test suite.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* remove names.MANAGER_NAMESPACE completely

after moving to use generalized managerNamespace
in cluster tests, we want to remove it completely from
tests, so we don't accidentally use it again.
This mean moving it from units tests as well

Signed-off-by: Ram Lavi <ralavi@redhat.com>