move 'test_id' tests to a new context, using the default opt-mode cluster config (#245)
* move test_id related tests to a new context

Currently all 'test_id' test are performed in opt-out context
There is no specific reason they were put there, since
they can run on either opt-mode, as long as they are running
on the same opt-mode configured by the cluster configuration (in order to prevent
the oprator deploying kubemacpool to reconcile back from the opt-mode configured by the test).
Let's therefor move the tests to a new context that will run using the
default opt-mode config that is set on the cluster.
Also this commit includes a small format fix in AllocateFillerVms

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* add BeforeEach to opt-In test

Since opt-mode contexts may run in different order,
we need to make sure that opt-in context start with
a "fresh" namespace.

Signed-off-by: Ram Lavi <ralavi@redhat.com>