logs, Make important logs more visible (#258)
Currently it is not easy to understand the the logs when in production mode
Also, we keep a 3 second reoccurring log on production which is annoying.
Refactored the logs in order to give a good picture of what's happening in
production mode.
Moved the 3 second reoccurring log to be a debug log.

Signed-off-by: Ram Lavi <ralavi@redhat.com>