vmToMacPoolMap, Add a vm name to interface:mac map
In order to follow the macs inserted/deleted from any vm we add a new
map to the macPool module.
The function implemented to update this map is updateVmToMacPoolMap().
We update added macs in the webhook context to avoid allocation races,
and update removed macs when vm is deleted.
We also populate it during the init of the KMP manager
Added vmToMacPoolMap to vm macpool module
Added updateVmToMacPoolMap() to handle mac additions/removals

Signed-off-by: Ram Lavi <ralavi@redhat.com>