ci, Allow for random conflict handling in tests
Occationally a webhook update request could get rejected due to
conlflict. These case are handled by design so that after a period of
time the rejected change is reverted.
Since our tests are also subjected to these conflicts, we should allow
for the kubemacpool cache to restore itself before going on with the
test.

Added retryOnConflictAndWaitForRestoreCache method that waits for the
configured waitTime should an update conflict occurr.
Changed all tests that update the vms to use this instead of the regular
retryOnConflict.
Added additional timeout to functest to allow for such conflict to be
handled.

Signed-off-by: Ram Lavi <ralavi@redhat.com>