webhook, Add kubemacpoolTransactionTimeStamp annotation
Introduce a new annotation to the virtualMachine instance:
kubemacpool.io/transaction-timestamp. This annotation will hold the timestamp
when the webhook runs.
The annotation format will be RFC3339Nano which allows smaller granolarity and
also accepts the old RFC3339 format if parsing old timestamps.
This annotation will be added by the KMP webhook, and updated everytime
a change is made to the vm spec.

Signed-off-by: Ram Lavi <ralavi@redhat.com>