pool, Change macPoolMap to hold transaction timestamp
In order to be able to properly update the macPoolMap, we expand it to
hold new parameters: the instance full name (vm/pod), macInstanceKey
name (interface name in vm case and network name in pod case), and
current transaction timestamp. The latter will replace the current
AllocationStatus field, where if timestamp is not set then the mac is
"no-pending-transaction".
The transactionTimestamp format is changed to RFC3339Nano which allows
smaller granolarity and also accepts the old RFC3339 format if parsing
old timestamps.
This commit only inserts the new methods related to the new macEntry,
along with method's unit-tests.

Signed-off-by: Ram Lavi <ralavi@redhat.com>