pool, Change macPoolMap to hold transaction timestamp
In order to be able to properly update the macPoolMap, we expand it to
hold new parameters: the instance full name (vm/pod), macInstanceKey
name (interface name in vm case and network name in pod case), and
current transaction timestamp. The latter will replace the current
AllocationStatus field.
The transactionTimestamp is saved in *time.Time type, where nil means
that the entry is not penging change, similar to former allocated status.
This commit only inserts the new methods related to the new macEntry,
along with method's unit-tests in a new mac-pool-map module.

Signed-off-by: Ram Lavi <ralavi@redhat.com>