pool, Change vm-allocate functions to support new macMap
This commit implements the new macEntry methods to functions allocating
the macs in vm webhook context and the macs setting the macs as ready on
the vm controller reconcile function.

In the webhook context, new macs are added to macPoolMap, and both
new/updated/deleted macs' entries get the latest transactionTimestamp.
These updated macs are later handled in the controller reconcile if
their transactionTimestamp is already due for update, compared to the
vm's
new kubemacpoolTransactionTimestamp annotation.

Also expanded and added to relevant unit tests,
Added a function checkMacPoolMapEntries that assists with checking macs
are properly allocated in the macPoolMap.

Signed-off-by: Ram Lavi <ralavi@redhat.com>