configMap, switch from using configMap to using the macPoolMap
As a result of expanding the macPoolMap, we can now discard using the
configMap and instead have vmWaitingCleanupLook goroutine go over the
macPoolMap to find stale entries.
For backwards compatibility, we still keep the configMap, and also keep
the legacy configMap cleaning mechanism, so that if during upgrade the
configMap is not empty, the mac's there will be incorporated to the
macPoolMap to prevent collisions.
Added e2e tests to check upgrade cases.
Also moved some of initVirtualMachineMap logic to a function
forEachVmInterfaceInClusterRunFunction in order to reuse it.

Signed-off-by: Ram Lavi <ralavi@redhat.com>