Change macPoolMap to use transaction timestamps (#269)
* ci, Add now as a global param

Replaces time.Now() to use a configurable func var stored in macPool
interface.
Currently there is not difference in behavior due to this change, but in
the future this will allow us to run unit tests with mocked time.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* webhook, Add kubemacpoolTransactionTimeStamp annotation

Introduce a new annotation to the virtualMachine instance:
kubemacpool.io/transaction-timestamp. This annotation will hold the timestamp
when the webhook runs.
The annotation format will be RFC3339Nano which allows smaller granolarity and
also accepts the old RFC3339 format if parsing old timestamps.
This annotation will be added by the KMP webhook, and updated everytime
a change is made to the vm spec.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool, Change macPoolMap to hold transaction timestamp

In order to be able to properly update the macPoolMap, we expand it to
hold new parameters: the instance full name (vm/pod), macInstanceKey
name (interface name in vm case and network name in pod case), and
current transaction timestamp. The latter will replace the current
AllocationStatus field.
The transactionTimestamp is saved in *time.Time type, where nil means
that the entry is not penging change, similar to former allocated status.
This commit only inserts the new methods related to the new macEntry,
along with method's unit-tests in a new mac-pool-map module.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool, Change vm-allocate functions to support new macMap

This commit implements the new macEntry methods to functions allocating
the macs in vm webhook context and the macs setting the macs as ready on
the vm controller reconcile function.

In the webhook context, new macs are added to macPoolMap, and both
new/updated/deleted macs' entries get the latest transactionTimestamp.
These updated macs are later handled in the controller reconcile if
their transactionTimestamp is already due for update, compared to the
vm's new kubemacpoolTransactionTimestamp annotation.

Also expanded and added to relevant unit tests,
Added a function checkMacPoolMapEntries that assists with checking macs
are properly allocated in the macPoolMap.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool, Change pod-allocate functions to support new macMap

This commit implements the new macEntry methods to functions allocating
the macs in pod webhook context and the macs setting the macs as ready
on the pod controller reconcile function.

In the webhook context, new macs are added to macPoolMap, and both new
macs' entries get the latest transactionTimestamp. These updated macs
are later handled in the controller reconcile.
Since pods are immutable and cannot be updated, we do not need to set
any transaction timestamp annotation on the pod like we do on the vm.
In case the pod's name is not yet updated when KMP webhook is invoked,
then we give it a temp name to prevent duplications

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* configMap, switch from using configMap to using the macPoolMap

As a result of expanding the macPoolMap, we can now discard using the
configMap and instead have vmWaitingCleanupLook goroutine go over the
macPoolMap to find stale entries.
For Backwards compatibility, we migrate the configMap at startup to the
macPoolMap, so the mac's there will be incorporated to the
macPoolMap to prevent collisions.
Added e2e tests to check upgrade cases.
Also moved some of initVirtualMachineMap logic to a function
forEachVmInterfaceInClusterRunFunction in order to reuse it.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* ci, Add e2e tests to check sequential vm updates

Added e2e tests that run many updates on the same vm to make sure the
last change is the one that persists on the macPoolMap.
Also increase e2e timeout to 1h.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* logs, reduce RequestId size

To make logs a bit shorter, reduce unique webhook/controller RequestId
to be up to 100000.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* ci, Add crash test after each vm e2e test

Signed-off-by: Ram Lavi <ralavi@redhat.com>