pod, Fix pod patching that causes the pod to error on creation (#274)
* pool, change network annotation var to public

We need the annotation in order to parse it during patching of the
webhook pod request.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pod, Fix pod patching that causes the pod to error on creation

Currently, a pod cannot be created when managed with Kubemacpool, due to
an error in the patching process of the webhook.
Fixed by patching the annotation needed.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pod-example, Remove issue warning

Now that the pod issue [1] is fixed, we can safely remove the Note from the
pod example.

[1] https://github.com/k8snetworkplumbingwg/kubemacpool/issues/273

Signed-off-by: Ram Lavi <ralavi@redhat.com>