webhook, vm, Include metadata changes to kubemacpool TS update scenario
Currently, only vm spec changes in the cause the kubemacpool
TransactionTimestampAnnotation to be updated. In order to optimize
updating of the mac to be faster, we also include metadata changes like
Annotation and label updates.
We do not include the TransactionTimestampAnnotation itself when
checking for changes, to avoid an internal update loop.

Signed-off-by: Ram Lavi <ralavi@redhat.com>