webhook, Include metadata changes to kubemacpool TS update scenario (#277)
* webhook, vm, Include metadata changes to kubemacpool TS update scenario

Currently, only vm spec changes in the cause the kubemacpool
TransactionTimestampAnnotation to be updated. In order to optimize
updating of the mac to be faster, we also include metadata changes like
Annotation and label updates.
We do not include the TransactionTimestampAnnotation itself when
checking for changes, to avoid an internal update loop.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* webhook, Fix Patching of annotations

Currently update of the annotation patch causes other existing
annotations to disappear, initiating a update loop with kubevirt trying
to add his annotation.
To avoid this, we move to patch the entire annotation field, similar to
how it is done in kubevirt[1].

[1]
https://github.com/kubevirt/kubevirt/blob/master/pkg/virt-api/webhooks/mutating-webhook/mutators/migration-create-mutator.go#L68

Signed-off-by: Ram Lavi <ralavi@redhat.com>