fix relaod if controller-manager.Start() returns
currently kubemacpool is waiting on signals in order to reload the
manager, simulating a kubemacpool restart.
An example is if kubevirt is not ready by the time kubemacpol is up.
In the example - when kubevirt is ready, a channel is closed to signal
kubemacpool to restart.
This commit fixes the waitForSignal method to use the new context based
cancel function.

Signed-off-by: Ram Lavi <ralavi@redhat.com>