finalizer, remove kubemacpool finalizer (#312)
we no longer need to keep the vm instance until kubemacpool has freed
the vm macs from the pool since we now store the vm-name in the macpool
mac entry.
Hence, we should remove the finalizer from the vm creation.
For backwards compatibility, we still remove the finalizer if it exists
when deleting the vm.

Signed-off-by: Ram Lavi <ralavi@redhat.com>