Filter the leader election pods by control-plane
The leader election mechanism tries to update a label on kubemacpool
pods to mark the one that is the leader but it search for all the pods
that are under namespace, this can end at massive number of pods
depending on which namespace is choose to install kubemacpool. This
change add a LabelSelector with control-plane: mac-controller-manager to
reduce the number of pods returned by the list operation.

Signed-off-by: Quique Llorente <ellorent@redhat.com>