dryRun, Adapt VM Mac Address allocation to support dryRun (#327)
* dryRun, Adapt VM Mac Address allocation to support dryRun

In case of "server" dryRun ("All" if used by controller runtime)
we do want to return an updated server response,
which allocates Mac Address, but without doing side effects.
Fix the flow to support it.

Note that the Mac pool head is changed, so calling twice dryRun
won't yield the same Mac Address in the response.
It is needed because in case there are multiple Mac Addresses in
the same request, we should not give the same Mac even on dryRun,
hence the mac pool head should be updated.

Signed-off-by: Or Shoval <oshoval@redhat.com>

* tests, Add vm dry run tests

Signed-off-by: Or Shoval <oshoval@redhat.com>

* dryRun, Adapt pod allocation to dryRun

Signed-off-by: Or Shoval <oshoval@redhat.com>