metrics: Use https instead of http (#346)
KMP uses controller runtime which creates an http
end point for /metrics.
In order to support https, we deploy rbac-proxy container
in the same pod of KMP.
The new container exposes https end point,
and connects internally to the http end point.

It is self signed, and protected with rbac rules.

Signed-off-by: Or Shoval <oshoval@redhat.com>