vm-pool, Reject VMs with duplicate interfaces
When applying a VM with multiple interfaces with the same name,
the VM is rejected but kubemacpool is mishandling the MACs,
causing a side effect where that MAC is no longer usable (it
is "taken" by the ghost VM).
This commit fixes this by checking and rejecting VMs with
duplicate interface names on kmp webhook context.

Signed-off-by: Ram Lavi <ralavi@redhat.com>