manager, Reduce toleration in case of node reboot
Currently if the node upon which KMP-manager deployment runs reboots,
the KMP pod is evicted only 5 minutes (default) after the node is unreachable.
Since KMP-manager deployment is is essential for VM create/update
service-ability, reducing this eviction to 1 minute. This is done by
altering the tolerations config [0]

This means that if the node is not available/ready for more than 1
minute, the pod will evict and will try to reschedule using the
configured affinity.

[0] https://kubernetes.io/docs/concepts/scheduling-eviction/taint-and-toleration/#taint-based-evictions

Signed-off-by: Ram Lavi <ralavi@redhat.com>