Fix duplicate check with different mac format (#397)
* Use removeMacEntry function instead of directly delete

macPoolMap provides a save remove entry function
that could always be preferred respect the
directly `delete` function.

Signed-off-by: fossedihelm <ffossemo@redhat.com>

* Fix duplicate check with different mac format

As per https://pkg.go.dev/net#ParseMAC, a macAddress can be specified
as an IEEE 802 MAC-48, EUI-48, EUI-64, or a 20-octet IP using the following formats:

- 00:00:5e:00:53:01
- 02:00:5e:10:00:00:00:01
- 00:00:00:00:fe:80:00:00:00:00:00:00:02:00:5e:10:00:00:00:01
- 00-00-5e-00-53-01
- 02-00-5e-10-00-00-00-01
- 00-00-00-00-fe-80-00-00-00-00-00-00-02-00-5e-10-00-00-00-01
- 0000.5e00.5301
- 0200.5e10.0000.0001
- 0000.0000.fe80.0000.0000.0000.0200.5e10.0000.0001

Currently, the duplicate mac check works only if the
two compared mac address use the same separator.
The duplicate check is made looking at the existence of key
of the macMap. The keys of the mapMap are the macAddress string.

With this commit, a `macKey` struct is introduced. The latter
stores the normalized string of the macAddress and can be used
as unique key and identifier of a macAddress.

NB: Currently KubeVirt does not allow EUI-64 and 20-octet IP mac addresses.

Signed-off-by: fossedihelm <ffossemo@redhat.com>

---------

Signed-off-by: fossedihelm <ffossemo@redhat.com>