pool_test: Increase pool range to reduce false positives
Since now MAC allocation is done randomly, it is better to increase the
MAC pool range to be large, to greatly reduce cases where the randomly
allocated MAC is the one accidentally expected by the tests.
Also change managedNamespaceMAC, unmanagedNamespaceMAC to be outside of
this range to reduce rare cases where the MAC allocator will assign that
MAC to the interfaces, creating a forbidden duplication.

Signed-off-by: Ram Lavi <ralavi@redhat.com>