webhook/vm: Fix transactionTS string not appearing sometimes
When the VM is reconciled but the change is not related to KMP (i.e.
interfaces, networks) then the TS is not created.
In this specific case the transactionTS is not created and because of
this - is overridden by an empty string.

Fix this in order to prevent the annotation from being overridden by an
empty string.

Signed-off-by: Ram Lavi <ralavi@redhat.com>