tests/virtual_machines: Fix transaction timestamp not persist on VM creation (#538)
* webhook/vm: Fix transactionTS string not appearing sometimes

When the VM is reconciled but the change is not related to KMP (i.e.
interfaces, networks) then the TS is not created.
In this specific case the transactionTS is not created and because of
this - is overridden by an empty string.

Fix this in order to prevent the annotation from being overridden by an
empty string.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* tests/virtual_machines: Add check for transaction timestamp

Add a check to make sure transaction timestamp is set on vm creation.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* webhook/vm: Refactor patch to accurately set annotation

Right now the annotation json patch adds the entire annotations map,
which could cause "reconciliation fights" with other controllers.

Moving to patching only the intended annotation.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

---------

Signed-off-by: Ram Lavi <ralavi@redhat.com>