pool-manager/pool_test: Assert expected MACs from vm's interfaces
Currently the unit test assumes that the first MAC in the range is the
one that is going to be allocated.
In preparation to changing the algorithm of the MAC allocation to not
start with the start of the range, changing the check to expect the MACs
set on the VM's interfaces, and only them.
This keeps the allocations logic and makes sure that the pool indeed
keeps the MACs handed to the VM's interfaces.

Signed-off-by: Ram Lavi <ralavi@redhat.com>