pool-manager: Refactor full ip pool e2e tests to unit test (#547)
* pool-mamager/pool: Fix documentation typo

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool-manager: Replace fullRange error by public var

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* pool-manager: Replace full range e2e by unit test

The scenario of full mac pool is not easy to recreate in a cluster with
a big range, and it could be easily replaced by a unit test.
Removing e2e tests that check this scenario.

Signed-off-by: Ram Lavi <ralavi@redhat.com>

* config/test: Remove small range patch

Currently the test manifests run on a small pool size of 16 MACs. This
was done in order to support checking e2e scenarios of full MAC pool
[0].
Now that these e2e test are removed, range in tests manifests can be
restored to the original values.

[0] https://github.com/k8snetworkplumbingwg/kubemacpool/pull/214

Signed-off-by: Ram Lavi <ralavi@redhat.com>

---------

Signed-off-by: Ram Lavi <ralavi@redhat.com>